Translate ideas and policies of HHReactor into AmpReactor
For the most part, Amp and Hack async-await are functionally equivalent,
which has allowed most of the fundamental workings of HHReactor to be
essentially directly translated into AmpReactor. There is a major
difference of the direction of control, although it's mostly behind the
scenes; where Hack's scheduler was hidden and felt "behind" the code as
it absorbed the await calls, the Amp scheduler looms above the async
code since we know we're yielding to it. Shaking that unfamiliarity
after learning more from Amp's source, there are more tangible
differences: notably, Deferred is most closely related to
ConditionWaitHandle, but with one crucial difference: the time bounds on
ConditionWaitHandle resolution restrict the scopes that can notify it,
whereas Deferred can stretch the entire application lifetime. They're
actually somewhat of closures around whatever is yield-awaiting them;
the system async call is callback based anyways, and this control flows
back to the iterators through these closures. They themselves then are
not "async" objects, but rather nicely, being bound (onResolve) by
Coroutine, Promise\wait and Producer only (if the application side stays
pull and doesn't bind a callback itself) they direct async control into
scopes where we expect async behavior.

MergeTest is failing because one of the clones' queues is missing the
first element consumed by the other clone. I knew I couldn't get away
with initializing the queue the same way, but reversing the order of
consumption and shifting from HHReactor; more on that in the next
commit.