Remove competition between two clones for first element in queue
If we clone an unstarted Producer, the first element won't be seen by
all the clones. This is from an oversight in the Producer
implementation: it was only checking that the buffer wasn't empty before
shifting, which works in the single-consumer case because that check
comes before `getCurrent` _can_ put any elements on the buffer. However,
with multiple consumers, the first to consume sees an empty queue but
none of the rest do, and those all miss the first element by shifting on
the first element.

Laziest solution: give them an element to shift on the first iteration.
It'll never be seen unless `getCurrent` is called before `advance` which
is undefined behavior anyways. If we need to support that, then we'll
have to add a "started" flag.