Add alpha group-by tests
Terming these "alpha" because they clone the subproducers to cool them
and compare items with the base array dimension instead of merging
across the 2nd dimension (to stick with the base-dimension testing
structure in `OperatorTestCase`).

This might however be misleading to the true behavior of group-by, where
the 2nd dimension producers are hot across both cold and hot consumers
of the base producer.