Add debounce tests and fix race condition on last element
Debounce forces the last element off the end of the lifetime of the base
producer because of the debouncing period. However, when this resolved,
the iterator was already complete and `Internal\Producer::emit` pushed
onto the queue that no-one would listen to (since `advance` had already
resolved to false once).

Instead, the internal time-delayed producer within `debounce` awaits the
final `Delayed`, _and_ one cycle of the event loop, in order to prevent a
race condition between that generator completing and the internal
consumer in `debounce` receiving and emitting the last element.